Genderファイルで値とメソッドの定義
