+++ RakutenRanking/RankingManager.swift
@@ -8,13 +8,6 @@
 
 import Foundation
 
-enum Gender: Int {
-    
-    case notKnown = 0
-    case male = 1
-    case female = 2
-}
-
 enum Age: Int {
     
     case notKnown = 0
